I Wrote The Cannonball Script
When a cannonball touches the character, the character dies. If it touches the platform, it explodes and if the character is inside the specified circle, he dies. I made the ball come out of the ball carriage in a parabola shape. If you are going to use the code, you can change the gun firing time, distance and height codes and use it.