The Character's Interaction With The Platforms Is Completed
The character has been enabled to stay on platforms with the rigidbody component, and now the character interacts with all platforms. So this means that if the platform rotates, the character will also rotate with the platform. If the platform is moving forward and there is a character on it, the character will move forward with the platform (without falling down).