First step for Character to Interact with Trap Door
In order for the character to interact with the trap door, I added an object with a z value of 3 on the trap door and turned off the mesh renderer component. I clicked the is trigger setting in the box collider and the character will be able to pass through this object.